Rename and fix RestrictGridSize
Should be called OverrideGridSize since it can be used to increase grid size
Since it is a debugging script, it now runs before our other MonoBehavior scripts.

The bug was that ChunkManager's Awake and GlobalDebugConfig's Awake had the same execution order.
Since GlobalDebugConfig is a debugging script, I deemed it valid to run it before our game's logic.